requirements.txt
aiohttp
coverage
matplotlib
numpy==2.3.5
pre-commit
pyjwt
pytz
requests
ruamel.yaml
